refactor: put LevelIndices in a module named after itself (it's big enough, I suppose)
